code/red_teaming_agent/model_scanning/requirements.txt
azure-ai-projects>=2.0.0
azure-ai-evaluation[redteam]
azure-identity
azure-ai-inference
openai>=1.0.0
